Make boundary checking ES5 compatible
The existing Regex uses property classes that were only introduced
in ES18 and thus can't be used in a polyfill. (They also don't
transpile nicely in Closure). This patch adds a script that uses
the regenerate and unicode-9.0.0 libraries to print (to console) a
suitable compatible replacement regex. I also replaced the
BOUNDARY_CHARS regex with the output from that script.