Adding check for whether or not range contains block boundary (#34)
This is the first step towards range (textStart + textEnd) matching: checking whether or not it's needed in the first place :)

The bulk of the logic here is just to check if a given range contains a block boundary (so that we can identify cases where an exact match is not possible). One test is currently disabled because of a known issue with forward traversal.