Using override map approach for traversal (#35)
Currently when trying to find block boundaries in the forward
direction, we use a postorder traversal so that the ends of blocks
can be found. However, this only makes sense if the traversal
started inside those nodes--otherwise, we'd want to hit the
leading edge. This change adopts a new approach based on
"injecting" these ancestor nodes into the traversal order using an
override map. This solves the known bugs (the disabled test).