Reorganizing FragmentFactory initialization
This patch adds the concept of modes to FragmentFactory, which
correspond to three new initialization methods which replace the
data formerly passed in at construct-time.

This is the last set-up step before adding context-based matches;
all that remains after this is to generate the search spaces and
pass them in.