Use Intl.Segmenter for highlighting, where available
This change supplies an Intl.Segmenter instance to the word bound
checking utils, where supported. This dramatically improves
the chances of a successful highlight in Japanese, Chinese, and
other non-space-separated languages.

This is not a breaking change. The old logic still exists and
is used in browsers which do not yet support Intl.Segmenter, and
the modified util functions treat the segmenter as an optional arg.
However, because this significantly affects real-world behavior, I
am treating this as a major version upgrade (4.0).