Add two util functions
Two new util functions are added to prepare for upcoming
Intl.Segmenter support in the fragment generation process. One
simply extracts the existing code for creating a Segmenter for
reuse. The other extracts all the text nodes at the same block
level as a given node, which will be needed in a forthcoming
patch.