Fix TODO in embiggen (#90)
This addresses two long-standing TODOs in the embiggen function.
The existing behavior does generate valid fragments, but not as
predictably as we'd like, because it doesn't account for the
possibility of multiple boundary characters in a row. This means
some iterations will not actually add a new word to the start/end.

Fixing this now better aligns the existing implementation with the
behavior of the Intl.Segmenter-based implementation, which will be
added as an option in a forthcoming patch.