Add more context/range words on first iteration
This updates a few behaviors in fragment generation to be more
consistent with the behavior of Chromium for Android and Desktop.

Specifically, the algorithm now:

* Adds 3 words on the first iteration and 1 word on subsequent
  iterations.
* Always adds context to exact text matches that are less than 20
  characters long.

A TODO also indicates where a future change will cause context to
be added more often to range-based matches.

These changes should help reduce the number of timeouts we see in
real-world behavior. The step of checking uniqueness can be quite
slow, especially if one or more params are short words that occur
many times in the document. By adding more words sooner, we have
a better chance of succeeding on the first attempt.