Adding test case for known issue with invisible text (#99)
When generating the search space for the prefix and suffix, we
attempt to exclude invisible nodes. However, invisible nodes
*between* visible nodes are incorrectly included in the search
space.

This patch adds a test case which is pending (disabled), to
facilitate development of a fix. A future patch will fix the bug
and enable the test.

This issue was first documented at crbug.com/1276014