This patch fixes a link generation bug.
Bug details: crbug.com/1289262

Fix: Before trying to generate a matching fragment, if the selected
range doesn't start or end in visible text (e.g: an image), we move
the edges of the range to the first and last visible text inside
the original selection. This removes the unexpected behaviors occurring
when the selection had non text content on its edges.