Test Infra: Added base class for data-driven tests.
This patch adds DataDrivenTest, a class used to implement data-driven
tests in the polyfill. This class includes most of the boilerplate
code needed to implement data-driven tests defined in input/ouput
files.

In a future PR, a subclass of DataDrivenTest will be used to implement
Text Fragment data-driven tests for the Test Infrastructure.