Fix findRangeFromNodeList
The function had a bug where it was using `range.startOffset`, which is
a raw (denormalized) offset, to index into a normalized string. This
could easily result in failure to find any matches, particularly when a
fragment had a prefix or a suffix. There were already three unit
tests failing because to this.

The bug seems to have been around for a long time, but it only started
manifesting after another fix was made last October:
https://github.com/GoogleChromeLabs/text-fragments-polyfill/commit/5d20c37450de6416f131b346041f5730af963a00

Before that fix, `range.startOffset` was never actually used, instead
the search always started from the beginning of the first text node.